# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trips
    description: Trip persistence and sync
  - name: profile
    description: Public profile (handle) for the signed-in user
  - name: discover
    description: Public Discover feed and public profile lookups
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips for the signed-in user
      description: Returns all trips that belong to the authenticated user.
      security:
        - clerkBearer: []
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
        "401":
          description: Missing or invalid auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: replaceTrips
      tags: [trips]
      summary: Replace all trips for the signed-in user
      description: >-
        Bulk-upserts the full list of trips for the authenticated user.
        Any trips previously stored for this user that are not in the body
        are deleted. This makes the endpoint idempotent and easy to call after
        every local change.
      security:
        - clerkBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trips]
              properties:
                trips:
                  type: array
                  items:
                    $ref: "#/components/schemas/Trip"
      responses:
        "200":
          description: Updated list of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get-or-create the signed-in user's public profile
      description: >-
        Returns the public profile for the authenticated user. If no profile
        exists yet, one is created on the fly with a handle derived from the
        user's primary email (everything to the left of `@`, lowercased,
        stripped of dots and `+tag`s, and with a numeric suffix if the
        derived handle is already taken). The handle is durable: it is set
        once at first call and not auto-rewritten if the user later changes
        their email.
      security:
        - clerkBearer: []
      responses:
        "200":
          description: My public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing or invalid auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /discover/trips:
    get:
      operationId: discoverTrips
      tags: [discover]
      summary: Public Discover feed of trips
      description: >-
        Cursor-paginated list of all trips that have been marked public,
        ordered newest-updated first. Open endpoint — no authentication
        required. Each item carries the trip and the author's public handle.
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor returned by a previous response
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
          description: Page size (default 20, max 50)
      responses:
        "200":
          description: Discover feed page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverFeed"
  /discover/trips/{id}:
    get:
      operationId: getPublicTrip
      tags: [discover]
      summary: Get a single public trip by id
      description: >-
        Returns one trip plus its author handle, but only if the trip has been
        marked public. 404 covers both "no such trip" and "trip exists but is
        private" so we never confirm to a stranger that a private trip exists.
        Open endpoint — no authentication required.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public trip detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverItem"
        "404":
          description: No such public trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{handle}:
    get:
      operationId: getPublicProfile
      tags: [discover]
      summary: Get a user's public profile and trips
      description: >-
        Returns the public profile and every public trip for the user with
        the given handle. Handle lookup is case-insensitive. Returns 404 if
        no user with this handle exists. Open endpoint — no authentication
        required.
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile and trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfilePage"
        "404":
          description: No user with that handle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued session JWT obtained via getToken() in the mobile client.
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Segment:
      type: object
      required:
        - id
        - fromCode
        - fromCity
        - toCode
        - toCity
        - depart
        - arrive
        - airline
        - flightNo
      properties:
        id:
          type: string
        fromCode:
          type: string
        fromCity:
          type: string
        toCode:
          type: string
        toCity:
          type: string
        depart:
          type: string
        arrive:
          type: string
        airline:
          type: string
        flightNo:
          type: string
        price:
          type: ["number", "null"]
    Branch:
      type: object
      required:
        - id
        - label
        - parentId
        - forkAfterSegmentId
        - color
        - segments
        - createdAt
      properties:
        id:
          type: string
        label:
          type: string
        parentId:
          type: ["string", "null"]
        forkAfterSegmentId:
          type: ["string", "null"]
        color:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"
        createdAt:
          type: string
    Trip:
      type: object
      required:
        - id
        - title
        - originCity
        - originCode
        - startDate
        - branches
        - activeBranchId
        - createdAt
      properties:
        id:
          type: string
        title:
          type: string
        originCity:
          type: string
        originCode:
          type: string
        startDate:
          type: string
        branches:
          type: array
          items:
            $ref: "#/components/schemas/Branch"
        activeBranchId:
          type: string
        createdAt:
          type: string
        isPublic:
          type: boolean
          description: When true, this trip is visible in the public Discover feed.
    UserProfile:
      type: object
      required: [handle, createdAt]
      properties:
        handle:
          type: string
          description: Public handle, derived from the user's email local part.
        createdAt:
          type: string
    PublicAuthor:
      type: object
      required: [handle]
      properties:
        handle:
          type: string
    DiscoverItem:
      type: object
      required: [trip, author, updatedAt]
      properties:
        trip:
          $ref: "#/components/schemas/Trip"
        author:
          $ref: "#/components/schemas/PublicAuthor"
        updatedAt:
          type: string
          description: Server timestamp of the last update to this trip.
    DiscoverFeed:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiscoverItem"
        nextCursor:
          type: ["string", "null"]
    PublicProfilePage:
      type: object
      required: [profile, trips]
      properties:
        profile:
          $ref: "#/components/schemas/UserProfile"
        trips:
          type: array
          items:
            $ref: "#/components/schemas/DiscoverItem"